+++ src/features/loadingSlice.ts
@@ -1,5 +1,5 @@
 import { PayloadAction, createSlice } from '@reduxjs/toolkit';
-import { RootState } from '../app/store';
+import { RootState } from '../store/store';
 
 type LoadingState = {
   main: boolean;
