+++ src/router/router.tsx
@@ -3,7 +3,7 @@ import {
   createRoutesFromElements,
   Route,
 } from 'react-router-dom';
-import App from '../App';
+import App from '../app/App';
 import ResultDetails from '../components/Results/ResultDetails/ResultDetails';
 
 export const router = createBrowserRouter(
